Control the RTS and DTR lines
I need to control the RTS and DTR lines to reset an ESP32 development board. But in general I think it's a good idea to have the control of the RTS and DTR lines implemented. I've implemented it as properties DtrEnable and RtsEnable like it's done in the the Windows interface.